feat: update team page layout and configuration
- Added output option to Next.js configuration for static export.
- Refactored team page to utilize tabs for better organization of tasks, activities, and members.
- Removed deprecated team detail page and updated sidebar links to use query parameters for team navigation.
- Enhanced user experience with improved component structure and state management.
