+++ app/teams/page.tsx
@@ -1,41 +1,60 @@
 'use client'
+import { TeamActivity } from '@/components/teams/TeamActivity'
+import { TeamMembers } from '@/components/teams/TeamMember'
+import { TeamDashboardHeader } from '@/components/teams/TeamDashboardHeader'
+import TeamTask from '@/components/teams/TeamTask'
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
+import { useState } from 'react'
+import { TeamTab } from '@/components/teams/TeamDashboardHeader'
+const tabs = [
+  {
+    name: 'Tasks',
+    value: 'tasks',
+    content: <TeamTask />,
+  },
+  {
+    name: 'Activities',
+    value: 'activities',
+    content: <TeamActivity />,
+  },
+  {
+    name: 'Members',
+    value: 'members',
+    content: <TeamMembers />,
+  },
+]
 
-import { Button } from '@/components/ui/button'
-import { useRouter } from 'next/navigation'
-import React from 'react'
-
-const TeamsPage = () => {
-  const router = useRouter()
+//TODO: we need to fetch the team details using the teamId we get in the query param
+export default function TeamPage() {
+  const [activeTab, setActiveTab] = useState(tabs[0].value)
 
   return (
-    <main className="flex min-h-[calc(100vh-4rem)] flex-1 items-start justify-center px-4 md:px-6 lg:px-8">
-      <div className="w-full max-w-sm text-center md:max-w-lg">
-        <h1 className="mb-3 px-2 text-xl font-bold md:mb-4 md:text-2xl lg:text-3xl">
-          Your Team space awaits!
-        </h1>
-        <h2 className="mb-6 px-2 text-sm font-semibold text-gray-500 md:mb-4 md:text-base lg:text-lg">
-          Collaborate better by joining or creating a team. Work together,
-          <br className="hidden sm:inline" />
-          assign tasks, and track progress as a unit
-        </h2>
+    <div className="flex pr-12">
+      <Tabs
+        defaultValue={tabs[0].value}
+        value={activeTab}
+        onValueChange={setActiveTab}
+        className="w-full px-12"
+      >
+        <TabsList className="bg-background w-full justify-start">
+          {tabs.map((tab) => (
+            <TabsTrigger
+              key={tab.value}
+              value={tab.value}
+              className="bg-background data-[state=active]:border-primary mr-6 h-full rounded-none border-b-2 border-transparent px-0 data-[state=active]:shadow-none"
+            >
+              <p className="text-left text-lg">{tab.name}</p>
+            </TabsTrigger>
+          ))}
+        </TabsList>
+        <TeamDashboardHeader activeTab={activeTab as TeamTab} />
 
-        <div className="mt-8 flex w-full flex-col items-center justify-center gap-3 px-4 md:mt-12 md:gap-4">
-          <Button
-            className="h-12 w-full max-w-xs cursor-pointer text-base md:h-auto md:text-sm"
-            onClick={() => router.push('/teams/create')}
-          >
-            Create a Team
-          </Button>
-          <Button
-            className="h-12 w-full max-w-xs cursor-pointer border border-black bg-white text-base font-bold text-black hover:bg-gray-50 md:h-auto md:text-sm"
-            onClick={() => router.push('/teams/join')}
-          >
-            Join an Existing Team
-          </Button>
-        </div>
-      </div>
-    </main>
+        {tabs.map((tab) => (
+          <TabsContent key={tab.value} value={tab.value} className="mt-4">
+            {tab.content}
+          </TabsContent>
+        ))}
+      </Tabs>
+    </div>
   )
 }
-
-export default TeamsPage
