+++ next.config.ts
@@ -2,6 +2,7 @@ import type { NextConfig } from 'next'
 
 const nextConfig: NextConfig = {
   /* config options here */
+  output: 'export',
   images: {
     unoptimized: true,
   },
